Rename SchedulingProfiling -> SchedulingProfiler
